Unify evidence bundle export in assembler
